requirements.txt
astreamlit
joblib
numpy
scikit-learn